Split buildDisplay into left/right files
