+++ Missionframework/modules/15_build/ui/buildDisplay_leftBg.hpp
@@ -1,0 +1,8 @@
+
+class KPLIB_DialogTitle: KPGUI_PRE_DialogTitle_LeftPanel {
+    text = "$STR_KPLIB_DIALOG_BUILD_TITLE";
+};
+
+class KP_DialogArea: KPGUI_PRE_DialogBackground_LeftPanel {
+    idc = KPLIB_IDC_BUILD_DIALOG_AREA;
+};
